+++ src/launcher/java/org/truffleruby/launcher/options/CommandLineParser.java
@@ -407,6 +407,7 @@ public class CommandLineParser {
                         // cancel other execution actions
                         config.setOption(OptionsCatalog.EXECUTION_ACTION, ExecutionAction.NONE);
                     } else if (extendedOption.startsWith("log=")) {
+                        LOGGER.warning("-Xlog= is deprecated and will be removed - use --log.ruby.level= instead");
                         final String level = extendedOption.substring("log=".length());
                         config.getUnknownArguments().add("--log.ruby.level=" + level);
                     } else {